Abstract server
