+++ integration/out-of-bounds-read.test.ts
@@ -1,7 +1,7 @@
 import { describe, before, test } from 'mocha'
 import { expect } from 'chai'
 import express from 'express'
-import { CreateHttpServer } from '../src/index'
+import { CreateHttpServer } from '../src/server'
 import {
   getOutOfBoundsReadCheri,
   getOutOfBoundsReadAarch64,
@@ -18,7 +18,7 @@ describe('Tests aarch64 version', () => {
   test('Happy Path - aarch64', async () => {
     const response = await getOutOfBoundsReadAarch64(app, 'pass', -32, -28)
 
-    console.log(response)
+    console.log(response.body)
     expect(response.status).to.equal(200)
     expect(response.body.status).to.contain('success')
   })
